Move control panel components to their own folder
