+++ client/src/components/edit/control-panel/ColorPicker.js
