+++ client/src/components/edit/control-panel/PixelSizeButtons.js
